requirements.txt
Flask==3.0.2
pandas==2.2.1
google-generativeai==0.4.1
PyPDF2==3.0.1
gunicorn==21.2.0